refactor: move services to separate dir for every app
