+++ analysis/filters.py
@@ -3,11 +3,11 @@ import logging
 import django_filters
 from django.db.models import Max, Q
 
-from patients.services import patient_service
+from patients.services.patients import patient_service
 from patients.errors import PatientServiceError
 from .models import VideoAnalysis
 
-logger = logging.getLogger(__name__)
+logger: logging.Logger = logging.getLogger(__name__)
 
 
 class VideoAnalysisFilter(django_filters.FilterSet):
@@ -110,7 +110,7 @@ class VideoAnalysisFilter(django_filters.FilterSet):
             return queryset
 
         try:
-            patients = patient_service.search_patients(search_term=value)
+            patients = patient_service.search_patients(pesel=value)
         except PatientServiceError as e:
             logger.error(f"Patient service error during search: {e}")
             raise
