+++ analysis/serializers.py
@@ -3,7 +3,6 @@ from rest_framework import serializers
 from drf_spectacular.utils import extend_schema_field
 from drf_spectacular.types import OpenApiTypes
 from .models import Substance, VideoAnalysis, AnalysisResult
-from patients.services import patient_service
 
 
 class SubstanceSerializer(serializers.ModelSerializer):
