+++ analysis/views/list_view.py
@@ -4,10 +4,10 @@ from django_filters.rest_framework import DjangoFilterBackend
 from analysis.models import VideoAnalysis
 from ..serializers import VideoAnalysisSerializer
 from ..filters import VideoAnalysisFilter
-from patients.services import patient_service
+from patients.services.patients import patient_service
 import logging
 
-logger = logging.getLogger(__name__)
+logger: logging.Logger = logging.getLogger(__name__)
 
 
 class VideoAnalysisListView(generics.ListCreateAPIView):
