+++ patients/views/get_patient_view.py
@@ -3,7 +3,7 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework import status
 
-from patients.services import patient_service
+from patients.services.patients import patient_service
 from patients.serializers import PatientSerializer
 from drf_spectacular.utils import extend_schema
